apply mild styling
