+++ client/src/components/Form/components/index.js
@@ -1,4 +1,5 @@
 import Actions from './Actions';
+import Autocomplete from './Autocomplete';
 import Content from './Content';
 import ContentText from './ContentText';
 import Input from './Input';
@@ -8,6 +9,7 @@ import { Select, MenuItem } from './Select';
 
 export {
     Actions,
+    Autocomplete,
     Content,
     ContentText,
     Input,
