+++ client/src/constant.js
@@ -1,3 +1,20 @@
+export const styles = {
+    active: {
+        backgroundColor: '#fafafa',
+        cursor: 'pointer',
+        padding: '7.5px 10px',
+        fontSize: 'small',
+        borderRadius: 4,
+    },
+    inactive: {
+        backgroundColor: '#ffffff',
+        cursor: 'pointer',
+        padding: '7.5px 10px',
+        fontSize: 'small',
+        borderRadius: 4,
+    },
+};
+
 export const defaultUsers = [
     {
         id: 0,
